object orient wrapper
