+++ tf_authoritative_scanner/util.py
@@ -15,3 +15,11 @@ def _read(rel_path):
     here = os.path.abspath(os.path.dirname(__file__))
     with codecs.open(os.path.join(here, rel_path), "r") as fp:
         return fp.read()
+
+
+def remove_leading_trailing_newline(text):
+    if text.startswith("\n"):
+        text = text[1:]
+    if text.endswith("\n"):
+        text = text[:-1]
+    return text
